[Modify] 토론 목록 공통 컴포넌트화
